Deletion, management, and tabcontrol
Sounds can now be right clicked and deleted, spent too much time formatting the tabcontrol. Currently it uses default microsoft code and needs to be changed to match the aesthetic of the program. Sounds now also play through the SoundManager, which fixed the problem that was causing sounds to suddenly stop playing. However, the memory for the sounds should still be released when they are done playing, I think.